feat(second-test): reorganize by prime version
When building in the same directory with a suffix, the common files are
overwrote with the latest versions things.  I'm not fully sure if this
is write so I was going to just merge the differences, but I think I
want to break the results by version so I can see them side by side.